- Logramos hacer la integración funcional - Ajustamos código para que funcione el dix de borrado de categorías en años diferentes al que esta actualmente abierto - El método findActualOpenYear deja de tener el asser que pedía que hubiera un año a lo sumo ya que se implementó código para detener cualquier tipo de scroll al abrir el selector de años como medida a evitar crash y ese sistema requiere preguntar por el año actualmente abierto

+++ Code/IncomesAndExpenses/IAEDateHelper.m
@@ -146,15 +146,15 @@
     const MonthType initialMonth = [[NSUserDefaults standardUserDefaults] actualInitialMonth];
     const BOOL usingTwoYears = initialMonth != January;
     if (usingTwoYears) {
-        NSString *firstYearString = [NSString stringWithFormat:@"%d", yearDate];
-        NSString *secondYearString = [NSString stringWithFormat:@"%d", yearDate + 1];
+        NSString *firstYearString = [NSString stringWithFormat:@"%lu", (unsigned long)yearDate];
+        NSString *secondYearString = [NSString stringWithFormat:@"%lu", yearDate + 1];
         if (shortForm) {
             firstYearString = [firstYearString substringWithRange:NSMakeRange(2, 2)];
             secondYearString = [secondYearString substringWithRange:NSMakeRange(2, 2)];
         }
         yearString = [NSString stringWithFormat:@"%@/%@", firstYearString, secondYearString];
     } else {
-        yearString = [NSString stringWithFormat:@"%d", yearDate];
+        yearString = [NSString stringWithFormat:@"%lu", (unsigned long)yearDate];
     }
     
     return yearString;
